Update README examples with template literals
Not sure if you count with some postprocessing of the JS, but with
```js
var html = '<extends src="base.html">' +
               '<block name="title" type="prepend">How to use posthtml-extend</block>' +
               '<block name="content">Read the documentation</block>' +
               '<block name="footer" type="append">— 2016</block>'
           '</extends>';
```

Variable `html` won't contain the last line. I've unified the examples to use the template literals with spacing and stuff like in the [example for plugins](https://github.com/posthtml/posthtml-extend#plugins).